Add tooltips to trade input
